optionB/B.1/StockPredictionPrice/B/1/v0.1/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
_tflow_select=2.2.0=eigen
abseil-cpp=20211102.0=hc377ac9_0
absl-py=2.1.0=py310hca03da5_0
aiohttp=3.9.5=py310h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.3=py310hca03da5_0
attrs=23.1.0=py310hca03da5_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=openblas
blinker=1.6.2=py310hca03da5_0
bottleneck=1.3.7=py310hbda83bc_0
brotli=1.0.9=h80987f9_8
brotli-bin=1.0.9=h80987f9_8
brotli-python=1.0.9=py310h313beb8_8
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2024.7.2=hca03da5_0
cachetools=5.3.3=py310hca03da5_0
certifi=2024.7.4=py310hca03da5_0
cffi=1.16.0=py310h80987f9_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py310hca03da5_0
contourpy=1.2.0=py310h48ca7d4_0
cryptography=41.0.3=py310h3c57c4d_0
cycler=0.11.0=pyhd3eb1b0_0
expat=2.6.2=h313beb8_0
flatbuffers=24.3.25=h313beb8_0
fonttools=4.51.0=py310h80987f9_0
freetype=2.12.1=h1192e45_0
frozendict=2.4.4=pypi_0
frozenlist=1.4.0=py310h80987f9_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h80987f9_3
google-auth=2.29.0=py310hca03da5_0
google-auth-oauthlib=0.5.2=py310hca03da5_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.48.2=h877324c_0
grpcio=1.48.2=py310h877324c_0
h5py=3.11.0=py310haafd478_0
hdf5=1.12.1=h160e8cb_2
html5lib=1.1=pypi_0
icu=68.1=hc377ac9_0
idna=3.7=py310hca03da5_0
joblib=1.4.2=py310hca03da5_0
jpeg=9e=h80987f9_3
keras=2.12.0=py310hca03da5_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=py310h313beb8_0
krb5=1.20.1=h8380606_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h80987f9_8
libbrotlidec=1.0.9=h80987f9_8
libbrotlienc=1.0.9=h80987f9_8
libcurl=8.2.1=h0f1d93c_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h80987f9_3
libnghttp2=1.52.0=h10c0552_1
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libssh2=1.10.0=h449679c_2
libtiff=4.5.1=h313beb8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h372ba2a_0
libxslt=1.1.37=habca612_0
llvm-openmp=14.0.6=hc6e5704_0
lxml=5.2.1=py310h50ffb84_0
lz4-c=1.9.4=h313beb8_1
markdown=3.4.1=py310hca03da5_0
markupsafe=2.1.3=py310h80987f9_0
matplotlib=3.8.4=py310hca03da5_0
matplotlib-base=3.8.4=py310h46d7db6_0
multidict=6.0.4=py310h80987f9_0
multitasking=0.0.11=pypi_0
ncurses=6.4=h313beb8_0
numexpr=2.8.7=py310hecc3335_0
numpy=1.23.5=py310hb93e574_0
numpy-base=1.23.5=py310haf87e8b_0
oauthlib=3.2.2=py310hca03da5_0
openjpeg=2.5.2=h54b8e55_0
openssl=1.1.1w=h1a28f6b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=24.1=py310hca03da5_0
pandas=2.2.2=py310h313beb8_0
pandas-datareader=0.10.0=pyhd3eb1b0_0
peewee=3.17.6=pypi_0
pillow=10.4.0=py310h80987f9_0
pip=24.0=py310hca03da5_0
platformdirs=4.2.2=pypi_0
protobuf=3.20.3=py310h313beb8_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pybind11-abi=4=hd3eb1b0_1
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.8.0=py310hca03da5_0
pyopenssl=23.2.0=py310hca03da5_0
pyparsing=3.0.9=py310hca03da5_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.13=hc0d8a6c_0
python-dateutil=2.9.0post0=py310hca03da5_2
python-flatbuffers=24.3.25=py310hca03da5_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310hca03da5_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
requests=2.32.3=py310hca03da5_0
requests-oauthlib=2.0.0=py310hca03da5_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.1.3=py310hc377ac9_0
scipy=1.13.1=py310hd336fd7_0
setuptools=72.1.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sklearn-pandas=2.2.0=py310hca03da5_0
snappy=1.1.10=h313beb8_1
soupsieve=2.5=pypi_0
sqlite=3.45.3=h80987f9_0
tensorboard=2.12.1=py310hca03da5_0
tensorboard-data-server=0.7.0=py310ha6e5c4f_0
tensorboard-plugin-wit=1.8.1=py310hca03da5_0
tensorflow=2.12.0=eigen_py310h205ab9b_0
tensorflow-base=2.12.0=eigen_py310h0a52ebb_0
tensorflow-estimator=2.12.0=py310hca03da5_0
termcolor=2.1.0=py310hca03da5_0
threadpoolctl=3.5.0=py310h33ce5c2_0
tk=8.6.14=h6ba3021_0
tornado=6.4.1=py310h80987f9_0
typing_extensions=4.11.0=py310hca03da5_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py310h80987f9_0
urllib3=2.2.2=py310hca03da5_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.3=py310hca03da5_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.14.1=py310h1a28f6b_0
xz=5.4.6=h80987f9_1
yarl=1.9.3=py310h80987f9_0
yfinance=0.2.41=pypi_0
zlib=1.2.13=h18a0788_1
zstd=1.5.5=hd90d995_2
